version 1.666 - fully supports custom encryption module for password lists

+++ encodelist.py
@@ -80,20 +80,27 @@ def gen_list():
          encfile = raw_input("invalid format. please enter a valid filename --> ")
       print("encoding each entry in password file..")
       encode_pass(origfile, encfile)
-      delsel = raw_input("delete the original unencoded password file? Y/N --> ")
+      delsel = raw_input("securely delete the original unencoded password file?\n ***YOU WILL NOT BE ABLE TO RECOVER THE FILE BECAUSE IT WILL BE REWRITTEN*** \nY/N --> ")
       while not re.match(r'^[nNyY]$', delsel):
-         delsel = raw_input("invalid selection. enter Y to delete or N to keep original unencoded file --> ")
+         delsel = raw_input("invalid selection. enter Y to securely delete or N to keep original unencoded file --> ")
       if delsel.lower() == 'y':
          try:
-            os.remove(origfile)
             if os.path.isfile(origfile):
-               os.unlink(origfile)
-            print('%s deleted successfully.' % origfile)
+               f = open(origfile, "wb")
+               f.write("*"*os.path.getsize(origfile))
+               f.close()
+               os.system('srm -m %s' % origfile)
+               if os.path.isfile(origfile):
+                  os.unlink(origfile)
+                  os.remove(origfile)
+               print('file %s securely removed successfully.' % origfile)
+            else:
+               print('file %s not found. nothing to delete.' % origfile)
          except OSError, e: 
             print ("Error: %s - %s." % (e.filename,e.strerror))
       else:
          if os.path.isfile(origfile):
-            print('%s contains sensitive data. please delete it manually ASAP.' % origfile)
+            print('%s may contain sensitive data. it is advised to securely delete it.' % origfile)
    else:
       encfile = raw_input("please enter filename of encoded password list --> ")
       while not os.path.exists(encfile):
@@ -104,21 +111,26 @@ def gen_list():
       print("decoding each entry in password file %s..") % encfile 
       encode_pass(encfile, newfile)
    def exitmenu():
-      exitsel = raw_input("enter 1 to run script again. enter 2 to run email2file script. enter 3 to print encoded/decoded data. to exit, enter 4 --> ")
+      exitsel = raw_input("enter 1 to run script again. \nenter 2 to run encryptlist.py script for extra security. \nenter 3 to run email2file script. \nenter 4 to print encoded/decoded data. \nto exit, enter 5 --> ")
 
-      while not re.search(r'^[1-4]$', exitsel):
-         exitsel = raw_input("invalid entry. enter 1 to run script again, 2 to run email2file script, 3 to show encoded/decoded data, or 4 to exit --> ")
+      while not re.search(r'^[1-5]$', exitsel):
+         exitsel = raw_input("invalid entry. enter 1 to run script again, 2 to run encryptlist.py, 3 to run email2file script, 4 to show encoded/decoded data, or 5 to exit --> ")
    
       if exitsel == '1':
          gen_list()
          exitmenu()
-   
+      
       elif exitsel == '2':
+         os.system('chmod +x encryptlist.py')
+         os.system('python encryptlist.py')
+         sys.exit()
+   
+      elif exitsel == '3':
          os.system('chmod +x email2file.py')
          os.system('./email2file.py')
          sys.exit()
    
-      elif exitsel == '3':
+      elif exitsel == '4':
          ef = open(encfile, "r+")
          for n in ef.readlines():
             print("encoded: %s" % n)
@@ -144,19 +156,19 @@ into a terminal under the same working directory.
  
  
 
-to run this script now and encrypt instead, enter 1.''')
+to continue with base64 encoding, enter 1.''')
 
-selscript = raw_input("to continue with base64 encoding, enter 2. --> ")
+selscript = raw_input("to switch to encryptlist.py instead, enter 2. --> ")
 
 while not re.match(r'^[12]$', selscript):
-   selscript = raw_input("invalid selection. to run encryption instead, enter 1. to continue with base64 encoding, enter 2. --> ")
+   selscript = raw_input("invalid selection. to continue with base64 encoding, enter 1. \nto run encryption module instead, enter 2. --> ")
    
 if selscript == '1':
-   os.system('python encryptlist.py')
-   sys.exit()
+   gen_list()
    
 else:
-   gen_list()
+   os.system('python encryptlist.py')
+   sys.exit()
 
 print("exiting program..")
